Fixed Trans Prob. Added Individual F parameter
The F parameter that we are inferring is the population relatedness. We are then taking 2 samples from this population and calculating the probability that those two individuals are related. This makes sense because we are in a fully bayesian framework and differs from the viterbri algorithm

At the moment K is being treated as a nuissance parameter